+++ Source/EasyNetQ/EasyNetQ.csproj
@@ -138,11 +138,6 @@
     <Compile Include="MessageVersioning\MessageVersionStack.cs" />
     <Compile Include="NonGeneric\NonGenericExtensions.cs" />
     <Compile Include="PersistentConnection.cs" />
-    <Compile Include="Producer\AdvancedClientRpc.cs" />
-    <Compile Include="Producer\BetterRpc.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Include="Producer\AdvancedServerRpc.cs" />
     <Compile Include="Producer\ClientCommandDispatcher.cs" />
     <Compile Include="Producer\ClientCommandDispatcherSingleton.cs" />
     <Compile Include="Producer\IClientCommandDispatcher.cs" />
@@ -180,6 +175,11 @@
     <Compile Include="Rpc\AdvancedRpc.cs" />
     <Compile Include="Rpc\IRpcHeaderKeys.cs" />
     <Compile Include="Rpc\RpcHeaderKeys.cs" />
+    <Compile Include="Rpc\AdvancedClientRpc.cs" />
+    <Compile Include="Rpc\AdvancedServerRpc.cs" />
+    <Compile Include="Rpc\BetterRpc.cs" />
+    <Compile Include="Rpc\IRpc.cs" />
+    <Compile Include="Rpc\Rpc.cs" />
     <Compile Include="Sprache\Failure.cs" />
     <Compile Include="Sprache\IFailure.cs" />
     <Compile Include="Sprache\Input.cs" />
