--- Source/EasyNetQ/Rpc/IAdvancedClientRpc.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-using EasyNetQ.Topology;
-
-namespace EasyNetQ.Rpc
-{
-    public interface IAdvancedClientRpc
-    {
-        Task<SerializedMessage> RequestAsync(IExchange requestExchange,
-                                             string requestRoutingKey,
-                                             bool mandatory,
-                                             bool immediate,
-                                             TimeSpan timeout,
-                                             SerializedMessage request);
-    }
-}
